Fix multiversion folder structure & Slightly adjust Gradle configuration
